added role-based access control

+++ app/Http/Controllers/JeloController.php
@@ -22,6 +22,10 @@ class JeloController extends Controller
 
     public function admin(Request $request):View
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         $jelos = Jelo::all();
 
         return view('jelo.admin', [
@@ -31,20 +35,32 @@ class JeloController extends Controller
 
     public function create(Request $request): View
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         return view('jelo.create');
     }
 
     public function store(JeloStoreRequest $request): RedirectResponse
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         $jelo = Jelo::create($request->validated());
 
         $request->session()->flash('jelo.id', $jelo->id);
 
-        return redirect()->route('jelos.index');
+        return redirect()->route('admin.jelo');
     }
 
     public function show(Request $request, Jelo $jelo): View
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         return view('jelo.show', [
             'jelo' => $jelo,
         ]);
@@ -52,6 +68,10 @@ class JeloController extends Controller
 
     public function edit(Request $request, Jelo $jelo): View
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         return view('jelo.edit', [
             'jelo' => $jelo,
         ]);
@@ -59,6 +79,10 @@ class JeloController extends Controller
 
     public function update(JeloUpdateRequest $request, Jelo $jelo): RedirectResponse
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         $jelo->update($request->validated());
 
         $request->session()->flash('jelo.id', $jelo->id);
@@ -68,8 +92,12 @@ class JeloController extends Controller
 
     public function destroy(Request $request, Jelo $jelo): RedirectResponse
     {
+        if (auth()->user()->role->naziv_role !== 'Admin')
+        {
+            abort(403, 'Nemate dozvolu za pristup.');
+        }
         $jelo->delete();
 
-        return redirect()->route('jelos.index');
+        return redirect()->route('admin.jelo');
     }
 }
